+++ mod.ts
@@ -6,8 +6,11 @@ export { InvalidTimeoutError, MissingApiKeyError } from "./src/errors.ts";
 export type {
   AccountApiParameters,
   AccountInformation,
+  AllowArbitraryParams,
   BaseParameters,
   BaseResponse,
+  EngineName,
+  EngineParameters,
   GetBySearchIdParameters,
   Location,
   Locations,